refactor: restructure api modules and entry file
