+++ index.html
@@ -453,6 +453,6 @@
         </div>
     </div>
 
-    <script type="module" src="src/gallery.js"></script>
+    <script type="module" src="src/index.js"></script>
 </body>
 </html>
